Add payment endpoint.
Payments are operations that are a part of executed transactions.  In
implementing this endpoint a few updates had to be made to the operation
resource.  First, the ID parameter of the operation is returned as a
string.  Our resources had IDs as integers as they appeared in the
stellar documentation.  Second, the serde_json::from_reader function
solely operates with owned data.  That means our operation deserializer
match
on a borrowed str was not compatible.  This PR also updates the
deserializer in operation to work with String.  We will need to
implement this in effects as well, but I will do this in a separate PR
when implementing an effect endpoint to keep the PR size manageable.